backend/requirements.txt
fastapi
uvicorn
ccxt>=4.0.0
pandas
numpy
python-dotenv
websockets
pydantic
httpx
requests
python-binance
